backend/requirements.txt
# Backend dependencies
flask
flask-cors
pymongo
python-dotenv
pyjwt
bcrypt
dnspython
requests
yt-dlp
